snowflake: support base64 encoded key
This is the same format that Snowflake's Kafka Connect driver takes and
I have seen snowflake savvy folks get tripped up, so let's fix that
papercut and support both formats.
